GraphAI-backend/backend/requirements.txt
# 核心框架
# 注意：cognee[neo4j] 0.5.1 需要 fastapi>=0.116.2 且 <1.0.0
fastapi>=0.116.2,<1.0.0
uvicorn[standard]>=0.24.0
# Graphiti 核心库（必须最先安装，让其他依赖自动解析）
graphiti-core==0.24.3
# Graphiti 的依赖要求（根据 graphiti-core 0.24.3 的要求）
# 重要：neo4j 必须在 cognee[neo4j] 之前安装，以保持版本 5.26.0
neo4j==5.26.0
pydantic>=2.11.5
openai>=1.91.0,<2.0.0
python-dotenv>=1.0.1
# 其他依赖（使用兼容版本）
pydantic-settings>=2.1.0
httpx>=0.25.2
numpy>=1.24.3
pandas>=2.1.3
python-multipart>=0.0.6
aiofiles>=23.2.1
python-docx>=0.8.11
openpyxl>=3.1.0
xlrd>=2.0.1
olefile>=0.46
# MySQL数据库
sqlalchemy>=2.0.0
pymysql>=1.1.0
cryptography>=41.0.0
# Celery和Redis
celery>=5.3.0
redis>=5.0.0
# JWT和密码加密
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
email-validator>=2.0.0
# Milvus向量数据库
pymilvus>=2.4.0
# Mem0 - 持久化记忆层
mem0ai>=1.0.0
ollama>=0.6.0
# Cognee - 章节知识图谱构建
# 注意：cognee[neo4j] 必须在 neo4j==5.26.0 之后安装，使用 --no-deps 避免升级 neo4j
# 此包将在 Dockerfile 中单独安装，这里仅作注释说明
# 固定使用 0.5.1 版本（0.5.2 在 add() 阶段使用 Instructor 导致 multiple tool calls 错误）
# cognee[neo4j]==0.5.1
# Cognee 的依赖（完整列表，确保所有依赖都安装）
structlog>=25.5.0
aiohttp>=3.11.14,<4.0.0
aiolimiter>=1.2.1
aiosqlite>=0.20.0,<1.0.0
alembic>=1.13.3,<2
fakeredis[lua]>=2.32.0
fastapi-users[sqlalchemy]>=14.0.1,<15.0.0
fastembed<=0.6.0
filetype>=1.2.0,<2.0.0
gunicorn>=20.1.0,<24
instructor>=1.9.1,<2.0.0
jinja2>=3.1.3,<4
kuzu==0.11.3
lancedb>=0.24.0,<1.0.0
limits>=4.4.1,<5
litellm>=1.76.0
mistralai>=1.9.10
nbformat>=5.7.0,<6.0.0
networkx>=3.4.2,<4
onnxruntime<=1.22.1
pympler>=1.1,<2.0.0
pypdf>=4.1.0,<7.0.0
rdflib>=7.1.4,<7.2.0
tiktoken>=0.8.0,<1.0.0
transformers>=4.30.0
# 注意：pydantic 和 websockets 版本冲突，需要在 Dockerfile 中处理
# pydantic<2.12.0,>=2.10.5 (cognee要求) vs pydantic>=2.11.5 (其他要求)
# websockets<16.0.0,>=15.0.1 (cognee要求) vs websockets 16.0 (其他要求)
# 注意：Milvus适配器可能需要单独安装或通过其他方式配置
